+++ src/state.rs
@@ -12,11 +12,11 @@ pub struct NfaState<'a> {
     state_params: HashMap<usize, StateParams>,
 }
 
-fn find_max<'a, I>(vals: I) -> f32
+fn find_max<'a, I>(vals: I) -> f64
 where
-    I: Iterator<Item = f32>,
+    I: Iterator<Item = f64>,
 {
-    vals.fold(f32::NEG_INFINITY, |a, b| a.max(b))
+    vals.fold(f64::NEG_INFINITY, |a, b| a.max(b))
 }
 
 impl NfaState<'_> {
@@ -29,14 +29,14 @@ impl NfaState<'_> {
         };
     }
 
-    pub fn update_states_counts(&mut self, states: &HashMap<usize, f32>) {
+    pub fn update_states_counts(&mut self, states: &HashMap<usize, f64>) {
         // TODO how is count dependent on p?
         states
             .into_iter()
             .for_each(|(i, p)| self.update_state(*i, 0.0, true));
     }
 
-    pub fn add_states(&mut self, states: &HashMap<usize, f32>, force: bool) -> f32 {
+    pub fn add_states(&mut self, states: &HashMap<usize, f64>, force: bool) -> f64 {
         find_max(
             states
                 .into_iter()
@@ -44,14 +44,14 @@ impl NfaState<'_> {
         )
     }
 
-    fn get_count(&self, idx: usize) -> u32 {
+    fn get_count(&self, idx: usize) -> u64 {
         if let Some(params) = self.state_params.get(&idx) {
             return params.2;
         }
         0
     }
 
-    fn get_state(&self, i: usize) -> (f32, &State) {
+    fn get_state(&self, i: usize) -> (f64, &State) {
         let p = match self.state_params.get(&i) {
             Some(params) => params.1,
             None => 0.0,
@@ -75,7 +75,7 @@ impl NfaState<'_> {
         }
     }
 
-    pub fn add_state(&mut self, idx: Option<usize>, force: bool, p: f32) -> f32 {
+    pub fn add_state(&mut self, idx: Option<usize>, force: bool, p: f64) -> f64 {
         let i = if let Some(i) = idx { i } else { return 0.0 };
 
         if let Some(state) = self.nfa.get(i) {
@@ -95,7 +95,7 @@ impl NfaState<'_> {
                 Kind::Quantifier(_) | Kind::Start | Kind::Split | Kind::ExactQuantifier(_) => {
                     let params = self.get_params(i);
                     let (p0, p1) = evaluate(p, Some(params));
-                    return f32::max(
+                    return f64::max(
                         self.add_state(state.outs.0, force, p0),
                         self.add_state(state.outs.1, force, p1),
                     );
@@ -113,15 +113,15 @@ impl NfaState<'_> {
         self.visited.drain();
     }
 
-    pub fn step(&mut self, token: char) -> f32 {
+    pub fn step(&mut self, token: char) -> f64 {
         debug!("step {}", token);
-        let mut new_states: HashMap<usize, f32> = HashMap::new();
+        let mut new_states: HashMap<usize, f64> = HashMap::new();
         let mut result = 0.0;
 
         let mut add_new_state = |idx, p| match idx {
             Some(i) => {
                 let e = new_states.entry(i).or_insert(0.0);
-                *e = f32::max(*e, p);
+                *e = f64::max(*e, p);
             }
             None => {}
         };
@@ -131,7 +131,7 @@ impl NfaState<'_> {
             match state.kind {
                 Kind::Terminal => {
                     debug!("  terminal");
-                    result = f32::max(p, result);
+                    result = f64::max(p, result);
                 }
                 Kind::Literal(match_token) => {
                     if match_token == token {
@@ -153,20 +153,20 @@ impl NfaState<'_> {
                 .map(|(k, v)| format!("p({})={}", k, v.1))
                 .join(" ")
         );
-        result = f32::max(result, self.add_states(&new_states, false));
+        result = f64::max(result, self.add_states(&new_states, false));
         self.visited.drain();
         debug!("  p terminal={}", result);
         return result;
     }
 
-    fn update_state(&mut self, i: usize, p: f32, count: bool) {
+    fn update_state(&mut self, i: usize, p: f64, count: bool) {
         let state = &self.nfa[i];
         debug!("      update {} p={} {}", state.kind, p, count);
 
         match state.kind {
             Kind::ExactQuantifier(n) => {
                 let mut params = self.get_params(i);
-                // params.1 = f32::max(params.1, p);
+                // params.1 = f64::max(params.1, p);
                 params.1 = p;
                 if count {
                     params.2 += 1;
@@ -175,7 +175,7 @@ impl NfaState<'_> {
             _ => {
                 self.current_states.insert(i);
                 let mut params = self.get_params(i);
-                // params.1 = f32::max(params.1, p);
+                // params.1 = f64::max(params.1, p);
                 params.1 = p;
                 if count {
                     params.2 += 1;
@@ -393,7 +393,7 @@ mod test {
             .keys()
             .sorted()
             .map(|k| state.state_params[k].1)
-            .collect::<Vec<f32>>();
+            .collect::<Vec<f64>>();
         assert_eq!(probs, vec![1.0, 0.0, 0.0]);
 
         assert_eq!(state.step('a'), 1.0);
@@ -402,7 +402,7 @@ mod test {
             .keys()
             .sorted()
             .map(|k| state.state_params[k].1)
-            .collect::<Vec<f32>>();
+            .collect::<Vec<f64>>();
         assert_eq!(probs, vec![0.0, 0.0, 1.0]);
     }
 }
